refactor one big file into a few modular files and clean a bit based on PEP
